+++ bill/bill.go
@@ -8,13 +8,15 @@ import (
 func init() {
 	schema.Register(schema.GOBL.Add("bill"),
 		Invoice{},
-		Receipt{},
+		Payment{},
 		CorrectionOptions{},
 	)
 }
 
 // Constants used to help identify document schemas
 const (
-	ShortSchemaInvoice = "bill/invoice"
-	ShortSchemaReceipt = "bill/receipt"
+	ShortSchemaOrder    = "bill/order"
+	ShortSchemaDelivery = "bill/delivery"
+	ShortSchemaInvoice  = "bill/invoice"
+	ShortSchemaPayment  = "bill/payment"
 )