+++ bill/payment.go
@@ -21,32 +21,32 @@ import (
 	"github.com/invopop/validation"
 )
 
-// Predefined list of the receipt types supported.
+// Predefined list of the payment types supported.
 const (
-	ReceiptTypePayment    cbc.Key = "payment"
-	ReceiptTypeRemittance cbc.Key = "remittance"
+	PaymentTypeReceipt cbc.Key = "receipt"
+	PaymentTypeAdvice  cbc.Key = "advice"
 )
 
-// ReceiptTypes defines the list of potential payment types.
-var ReceiptTypes = []*cbc.Definition{
+// PaymentTypes defines the list of potential payment types.
+var PaymentTypes = []*cbc.Definition{
 	{
-		Key: ReceiptTypePayment,
+		Key: PaymentTypeReceipt,
 		Name: i18n.String{
-			i18n.EN: "Payment",
+			i18n.EN: "Receipt",
 		},
 		Desc: i18n.String{
 			i18n.EN: here.Doc(`
 				A payment receipt sent from the supplier to a customer indicating that they have
 				received a transfer of funds from the customer directly or a payer.
-				This is the default receipt type and may be required by some tax
+				This is the default payment type and may be required by some tax
 				regimes in order to communicate the payment of specific documents and invoices.
 			`),
 		},
 	},
 	{
-		Key: ReceiptTypeRemittance,
+		Key: PaymentTypeAdvice,
 		Name: i18n.String{
-			i18n.EN: "Remittance",
+			i18n.EN: "Advice",
 		},
 		Desc: i18n.String{
 			i18n.EN: here.Doc(`
@@ -57,24 +57,24 @@ var ReceiptTypes = []*cbc.Definition{
 	},
 }
 
-var isValidReceiptType = validation.In(validReceiptTypes()...)
+var isValidPaymentType = validation.In(validPaymentTypes()...)
 
-func validReceiptTypes() []interface{} {
-	list := make([]interface{}, len(ReceiptTypes))
-	for i, d := range ReceiptTypes {
+func validPaymentTypes() []interface{} {
+	list := make([]interface{}, len(PaymentTypes))
+	for i, d := range PaymentTypes {
 		list[i] = d.Key
 	}
 	return list
 }
 
-// A Receipt is used to link an invoice or invoices with a payment transaction.
-type Receipt struct {
+// A Payment is used to link an invoice or invoices with a payment transaction.
+type Payment struct {
 	tax.Regime
 	tax.Addons
 	tax.Tags
 	uuid.Identify
 
-	// Type of receipt document being issued.
+	// Type of payment document being issued.
 	Type cbc.Key `json:"type" jsonschema:"title=Type" jsonschema_extras:"calculated=true"`
 	// Details on how the payment was made based on the original instructions.
 	Method *pay.Instructions `json:"method,omitempty" jsonschema:"title=Method"`
@@ -104,7 +104,7 @@ type Receipt struct {
 	Payee *org.Party `json:"payee,omitempty" jsonschema:"title=Payee"`
 
 	// List of documents that are being paid for.
-	Lines []*ReceiptLine `json:"lines" jsonschema:"title=Lines"`
+	Lines []*PaymentLine `json:"lines" jsonschema:"title=Lines"`
 
 	// Ordering allows for additional information about the ordering process including references
 	// to other documents and alternative parties involved in the order-to-delivery process.
@@ -113,77 +113,77 @@ type Receipt struct {
 	// Summary of the taxes applied to the payment for tax regimes that require
 	// this information to be communicated.
 	Tax *tax.Total `json:"tax,omitempty" jsonschema:"title=Tax"`
-	// Total amount to be paid in this receipt, either positive or negative according to the
+	// Total amount to be paid in this payment, either positive or negative according to the
 	// line types and totals.
 	Total num.Amount `json:"total" jsonschema:"title=Total"`
 
-	// Unstructured information that is relevant to the receipt, such as correction or additional
+	// Unstructured information that is relevant to the payment, such as correction or additional
 	// legal details.
 	Notes []*org.Note `json:"notes,omitempty" jsonschema:"title=Notes"`
 
-	// Additional complementary objects that add relevant information to the receipt.
+	// Additional complementary objects that add relevant information to the payment.
 	Complements []*schema.Object `json:"complements,omitempty" jsonschema:"title=Complements"`
 
 	// Additional semi-structured data that doesn't fit into the body of the invoice.
 	Meta cbc.Meta `json:"meta,omitempty" jsonschema:"title=Meta"`
 }
 
-// Validate runs the validation rules for the receipt without the context.
-func (rct *Receipt) Validate() error {
-	return rct.ValidateWithContext(context.Background())
+// Validate runs the validation rules for the payment without the context.
+func (pmt *Payment) Validate() error {
+	return pmt.ValidateWithContext(context.Background())
 }
 
-// ValidateWithContext ensures that the fields contained in the Receipt look correct.
-func (rct *Receipt) ValidateWithContext(ctx context.Context) error {
-	ctx = rct.validationContext(ctx)
-	r := rct.RegimeDef()
-	return tax.ValidateStructWithContext(ctx, rct,
-		validation.Field(&rct.Regime),
-		validation.Field(&rct.Addons),
-		validation.Field(&rct.UUID),
-		validation.Field(&rct.Type,
+// ValidateWithContext ensures that the fields contained in the Payment look correct.
+func (pmt *Payment) ValidateWithContext(ctx context.Context) error {
+	ctx = pmt.validationContext(ctx)
+	r := pmt.RegimeDef()
+	return tax.ValidateStructWithContext(ctx, pmt,
+		validation.Field(&pmt.Regime),
+		validation.Field(&pmt.Addons),
+		validation.Field(&pmt.UUID),
+		validation.Field(&pmt.Type,
 			validation.Required,
-			isValidReceiptType,
+			isValidPaymentType,
 		),
-		validation.Field(&rct.Method, validation.Required),
-		validation.Field(&rct.Series),
-		validation.Field(&rct.Code,
+		validation.Field(&pmt.Method, validation.Required),
+		validation.Field(&pmt.Series),
+		validation.Field(&pmt.Code,
 			validation.When(
 				internal.IsSigned(ctx),
-				validation.Required.Error("required to sign receipt"),
+				validation.Required.Error("required to sign payment"),
 			),
 		),
-		validation.Field(&rct.IssueDate,
+		validation.Field(&pmt.IssueDate,
 			validation.Required,
 			cal.DateNotZero(),
 		),
-		validation.Field(&rct.Currency,
+		validation.Field(&pmt.Currency,
 			validation.Required,
-			currency.CanConvertInto(rct.ExchangeRates, r.GetCurrency()),
+			currency.CanConvertInto(pmt.ExchangeRates, r.GetCurrency()),
 		),
-		validation.Field(&rct.ExchangeRates),
-		validation.Field(&rct.Ext),
-		validation.Field(&rct.Preceding),
-		validation.Field(&rct.Supplier, validation.Required),
-		validation.Field(&rct.Customer),
-		validation.Field(&rct.Payee),
-		validation.Field(&rct.Lines, validation.Required),
-		validation.Field(&rct.Ordering),
-		validation.Field(&rct.Tax),
-		validation.Field(&rct.Total, validation.Required),
-		validation.Field(&rct.Notes),
-		validation.Field(&rct.Complements),
-		validation.Field(&rct.Meta),
+		validation.Field(&pmt.ExchangeRates),
+		validation.Field(&pmt.Ext),
+		validation.Field(&pmt.Preceding),
+		validation.Field(&pmt.Supplier, validation.Required),
+		validation.Field(&pmt.Customer),
+		validation.Field(&pmt.Payee),
+		validation.Field(&pmt.Lines, validation.Required),
+		validation.Field(&pmt.Ordering),
+		validation.Field(&pmt.Tax),
+		validation.Field(&pmt.Total, validation.Required),
+		validation.Field(&pmt.Notes),
+		validation.Field(&pmt.Complements),
+		validation.Field(&pmt.Meta),
 	)
 }
 
-// validationContext builds a context with all the validators that the receipt might
+// validationContext builds a context with all the validators that the payment might
 // need for execution.
-func (rct *Receipt) validationContext(ctx context.Context) context.Context {
-	if r := rct.RegimeDef(); r != nil {
+func (pmt *Payment) validationContext(ctx context.Context) context.Context {
+	if r := pmt.RegimeDef(); r != nil {
 		ctx = r.WithContext(ctx)
 	}
-	for _, a := range rct.AddonDefs() {
+	for _, a := range pmt.AddonDefs() {
 		ctx = a.WithContext(ctx)
 	}
 	return ctx
@@ -192,66 +192,66 @@ func (rct *Receipt) validationContext(ctx context.Context) context.Context {
 // Calculate performs all the normalizations and calculations required for the invoice
 // totals and taxes. If the original invoice only includes partial calculations, this
 // will figure out what's missing.
-func (rct *Receipt) Calculate() error {
+func (pmt *Payment) Calculate() error {
 	// Try to set Regime if not already prepared from the supplier's tax ID
-	if rct.Regime.IsEmpty() {
-		rct.SetRegime(partyTaxCountry(rct.Supplier))
+	if pmt.Regime.IsEmpty() {
+		pmt.SetRegime(partyTaxCountry(pmt.Supplier))
 	}
-	rct.Normalize(rct.normalizers())
-	return rct.calculate()
+	pmt.Normalize(pmt.normalizers())
+	return pmt.calculate()
 }
 
 // Normalize is run as part of the Calculate method to ensure that the invoice
 // is in a consistent state before calculations are performed. This will leverage
 // any add-ons alongside the tax regime.
-func (rct *Receipt) Normalize(normalizers tax.Normalizers) {
-	if rct.Type == cbc.KeyEmpty {
-		rct.Type = ReceiptTypePayment
+func (pmt *Payment) Normalize(normalizers tax.Normalizers) {
+	if pmt.Type == cbc.KeyEmpty {
+		pmt.Type = PaymentTypeReceipt
 	}
-	rct.Series = cbc.NormalizeCode(rct.Series)
-	rct.Code = cbc.NormalizeCode(rct.Code)
+	pmt.Series = cbc.NormalizeCode(pmt.Series)
+	pmt.Code = cbc.NormalizeCode(pmt.Code)
 
-	normalizers.Each(rct)
+	normalizers.Each(pmt)
 
-	tax.Normalize(normalizers, rct.Method)
-	tax.Normalize(normalizers, rct.Tax)
-	tax.Normalize(normalizers, rct.Supplier)
-	tax.Normalize(normalizers, rct.Customer)
-	tax.Normalize(normalizers, rct.Preceding)
-	tax.Normalize(normalizers, rct.Lines)
-	tax.Normalize(normalizers, rct.Ordering)
+	tax.Normalize(normalizers, pmt.Method)
+	tax.Normalize(normalizers, pmt.Tax)
+	tax.Normalize(normalizers, pmt.Supplier)
+	tax.Normalize(normalizers, pmt.Customer)
+	tax.Normalize(normalizers, pmt.Preceding)
+	tax.Normalize(normalizers, pmt.Lines)
+	tax.Normalize(normalizers, pmt.Ordering)
 }
 
-func (rct *Receipt) normalizers() tax.Normalizers {
+func (pmt *Payment) normalizers() tax.Normalizers {
 	normalizers := make(tax.Normalizers, 0)
-	if r := rct.RegimeDef(); r != nil {
+	if r := pmt.RegimeDef(); r != nil {
 		normalizers = normalizers.Append(r.Normalizer)
 	}
-	for _, a := range rct.AddonDefs() {
+	for _, a := range pmt.AddonDefs() {
 		normalizers = normalizers.Append(a.Normalizer)
 	}
 	return normalizers
 }
 
-func (rct *Receipt) calculate() error {
+func (pmt *Payment) calculate() error {
 	var tt *tax.Total
 	var total *num.Amount
 
-	r := rct.RegimeDef()
+	r := pmt.RegimeDef()
 
 	// Convert empty or invalid currency to the regime's currency
-	if rct.Currency == currency.CodeEmpty && r != nil {
-		rct.Currency = r.Currency
+	if pmt.Currency == currency.CodeEmpty && r != nil {
+		pmt.Currency = r.Currency
 	}
-	if rct.Currency == currency.CodeEmpty {
+	if pmt.Currency == currency.CodeEmpty {
 		return validation.Errors{
 			"currency": fmt.Errorf("required, unable to determine"),
 		}
 	}
 
-	for i, l := range rct.Lines {
+	for i, l := range pmt.Lines {
 		l.Index = i + 1
-		if err := l.calculate(rct.Currency, rct.ExchangeRates); err != nil {
+		if err := l.calculate(pmt.Currency, pmt.ExchangeRates); err != nil {
 			return validation.Errors{
 				"lines": validation.Errors{
 					strconv.Itoa(i): err,
@@ -259,7 +259,7 @@ func (rct *Receipt) calculate() error {
 			}
 		}
 
-		l.Tax.Calculate(rct.Currency, r.GetRoundingRule())
+		l.Tax.Calculate(pmt.Currency, r.GetRoundingRule())
 		lt := l.Tax.Clone()
 		a := l.Total
 
@@ -280,20 +280,20 @@ func (rct *Receipt) calculate() error {
 			total = &nt
 		}
 	}
-	rct.Tax = tt
+	pmt.Tax = tt
 	if total != nil {
-		rct.Total = *total
+		pmt.Total = *total
 	}
 	return nil
 }
 
 // JSONSchemaExtend extends the schema with additional property details
-func (rct Receipt) JSONSchemaExtend(js *jsonschema.Schema) {
+func (pmt Payment) JSONSchemaExtend(js *jsonschema.Schema) {
 	props := js.Properties
 	// Extend type list
 	if its, ok := props.Get("type"); ok {
-		its.OneOf = make([]*jsonschema.Schema, len(ReceiptTypes))
-		for i, kd := range ReceiptTypes {
+		its.OneOf = make([]*jsonschema.Schema, len(PaymentTypes))
+		for i, kd := range PaymentTypes {
 			its.OneOf[i] = &jsonschema.Schema{
 				Const:       kd.Key.String(),
 				Title:       kd.Name.String(),
@@ -301,8 +301,8 @@ func (rct Receipt) JSONSchemaExtend(js *jsonschema.Schema) {
 			}
 		}
 	}
-	rct.Regime.JSONSchemaExtend(js)
-	rct.Addons.JSONSchemaExtend(js)
+	pmt.Regime.JSONSchemaExtend(js)
+	pmt.Addons.JSONSchemaExtend(js)
 	// Recommendations
 	js.Extras = map[string]any{
 		schema.Recommended: []string{